Make meeples exit the game upon escaping.
This introduces a new database column to track which meeples have already
left. Also there is a set of new controls in the lower left signalling this
status. These controls will be used in the future to easier locate the
individual meeples.

This also changes some internal logic, for example the check for the win
condition.